Mark

+++ src/Console/Commands/Game/DataSyncCommand.php
@@ -533,7 +533,7 @@ class DataSyncCommand extends Command implements CommandInterface
                         'scope' => object_get($spellData, 'scope'),
                         'class' => object_get($spellData, 'class'),
                         'cost' => object_get($spellData, 'cost', 1),
-                        'duration' => object_get($spellData, 'duration', 0),
+                        'duration' => (float)object_get($spellData, 'duration', 0),
                         'cooldown' => object_get($spellData, 'cooldown', 0),
                         'wizard_strength' => object_get($spellData, 'wizard_strength'),
                         'deity_id' => $deityId,
